+++ server/src/utils/serverDependencies.ts
@@ -53,7 +53,6 @@ interface RouteModules {
   shotspotterZonesRoutes: Router;
   shotspotterSensorsRoutes: Router;
   networkAgenciesRoutes: Router;
-  siblingDetectionRoutes: Router;
 }
 
 /**
@@ -113,7 +112,6 @@ function loadRouteModules(): RouteModules {
     shotspotterZonesRoutes: require('../api/routes/v1/shotspotterZones').default,
     shotspotterSensorsRoutes: require('../api/routes/v1/shotspotterSensors').default,
     networkAgenciesRoutes: require('../api/routes/v1/network-agencies'),
-    siblingDetectionRoutes: require('../api/routes/v1/siblingDetection'),
   };
 }